requirements.txt
# this is a file containing needed packages for project */
# simply run from commandline: 
# py -3 -m pip install -r /path/to/requirements.txt 
requests
social-auth-app-django
python-social-auth[django]
pyjwkest
tweepy
Redis
django-qrauth
//qrauth lib has been modified to work with Django 1.11. It has been moved to project folder;